add button for looking at next maximum
next maximum is basically any clique that of size greater than the lower
bound, and that does NOT contain any part of the existing clique. Such a
setup is currently a destructive process -- the graph amat will be
pruned continually as you keep asking for the next clique.

this can be tested by setting the distortion values to 1 and keeping the
lower bound as 5 or 6, so as to get spurious maxima.